Implement `or.b imm, #(r0, gbr)`; refactor MA stage
The `or.b ...` instruction executes EX MA EX MA, and stalls the pipeline
twice. This required changing the stages to allow a stage to be executed
more than once, so I added `Stage::step`.

The MA stage needs to know when it should proceed to the WB stage, and
when to go back to the EX stage, so I added a few more helper functions:

* `stage_ma_read_u{8,16,32}_ex`
* `stage_ma_read_u{8,16,32}_wb`

The `*_wb` functions are renames of the existing functions.

This fix means that I can remove the
`STAGE_RESULT_WRITEBACK_AS_EXECUTE` hack, since `lcd.l *` can actually
go back to the EX stage now.